Add files via upload
The PDFs are in a main folder named "pdf", containing each files are named accordingly to the name of the subjects. computer, Gk, history, science_fiction respectively.
The images are in a folder named "img".
Style sheet files are in "CSS" folder.
Java script file is in js folder.

NOTE THAT: HERE IN github no folders can be uploaded. So I have uploaded only the files. please make sure to create folders containing the files accordingly mentioned above, for the appropriate output.  
Thank you.